+++ extras/tests/JsonArray/subscript.cpp
@@ -135,7 +135,7 @@ TEST_CASE("JsonArray::operator[]") {
 
 #ifdef HAS_VARIABLE_LENGTH_ARRAY
   SECTION("set(VLA)") {
-    int i = 16;
+    size_t i = 16;
     char vla[i];
     strcpy(vla, "world");
 
@@ -146,7 +146,7 @@ TEST_CASE("JsonArray::operator[]") {
   }
 
   SECTION("operator=(VLA)") {
-    int i = 16;
+    size_t i = 16;
     char vla[i];
     strcpy(vla, "world");
