+++ extras/tests/MsgPackDeserializer/filter.cpp
@@ -1068,7 +1068,7 @@ TEST_CASE("Overloads") {
 
 #ifdef HAS_VARIABLE_LENGTH_ARRAY
   SECTION("char[n], Filter") {
-    int i = 4;
+    size_t i = 4;
     char vla[i];
     strcpy(vla, "{}");
     deserializeMsgPack(doc, vla, Filter(filter));
@@ -1096,7 +1096,7 @@ TEST_CASE("Overloads") {
 
 #ifdef HAS_VARIABLE_LENGTH_ARRAY
   SECTION("char[n], Filter, NestingLimit") {
-    int i = 4;
+    size_t i = 4;
     char vla[i];
     strcpy(vla, "{}");
     deserializeMsgPack(doc, vla, Filter(filter), NestingLimit(5));
@@ -1124,7 +1124,7 @@ TEST_CASE("Overloads") {
 
 #ifdef HAS_VARIABLE_LENGTH_ARRAY
   SECTION("char[n], NestingLimit, Filter") {
-    int i = 4;
+    size_t i = 4;
     char vla[i];
     strcpy(vla, "{}");
     deserializeMsgPack(doc, vla, NestingLimit(5), Filter(filter));
