+++ extras/tests/MsgPackDeserializer/input_types.cpp
@@ -72,7 +72,7 @@ TEST_CASE("deserializeMsgPack(std::istream&)") {
 
 #ifdef HAS_VARIABLE_LENGTH_ARRAY
 TEST_CASE("deserializeMsgPack(VLA)") {
-  int i = 16;
+  size_t i = 16;
   char vla[i];
   memcpy(vla, "\xDE\x00\x01\xA5Hello\xA5world", 15);
